* commit 'ca604a5f43f569c3fb06311dedd50adb50fae291':
  SW-22779 - Update bi version

+++ engine/Shopware/Bundle/BenchmarkBundle/Provider/ShipmentsProvider.php
@@ -100,8 +100,8 @@ class ShipmentsProvider implements BenchmarkProviderInterface
         }
 
         $matches['others']['name'] = 'others';
-        $matches['others']['minPrice'] = min(array_column($others, 'minPrice'));
-        $matches['others']['maxPrice'] = max(array_column($others, 'maxPrice'));
+        $matches['others']['minPrice'] = (float) min(array_column($others, 'minPrice'));
+        $matches['others']['maxPrice'] = (float) max(array_column($others, 'maxPrice'));
 
         return array_values($matches);
     }
